Update login component CSS and HTML; replace logo image
The login page's appearance has been modified by updating the CSS of the login component. The background color of the left container and text container has been changed to #1E1E1E. The height of the logo and formInput has been adjusted. The logo image has been replaced in both the CSS and HTML files with a new image file, "OnlineOrderingImage-removebg-preview.png".