Changed backend User's model from an array of restaurants to a singular restaurant.
Along with the Restaurant model from an array of Owners to a singular Owner.

Updated all methods on backend and front end that go along with these model changed.

Added and styled a 'addAdminDialog' to allow the creation of an 'Admin"' as in owner of the restaurant to be created.

Added a new type to be able to create a user without a restaurant object since the user will be created before the restaurant in this specific scenario.

Will update functionality of these and finish the add restaurant capabilites.